(dev) fifo fixed posix semantics and races

+++ userspace/lib/include/unit.h
@@ -46,7 +46,7 @@ UnitID spawn_unit(RealmID realm_id, uint64_t entry_point, uint64_t arg_ptr, uint
  * @param exit_code_out Pointer to store the unit's exit/return code, or NULL if unused.
  * @return 0 on success, negative error code on failure (e.g. -ESRCH if unit_id invalid/already reaped).
  */
-int64_t join_unit(UnitID unit_id, int64_t* exit_code_out);
+int64_t join_unit(UnitID unit_id, int* exit_code_out);
 
 
 UnitID get_unit_id(void);
